Restructure plugin package
- Split plugin.ts up into different files
- Fix pluginPath being used instead of modulePath in Plugin:loadModule
- Fix repository link in package.json
