+++ packages/plugins/lib/index.ts
@@ -1,1 +1,4 @@
-export * from "./plugin";
+export * from "./interfaces";
+export * from "./Plugin";
+export * from "./Plugins";
+export * from "./util";
